Updated ability to show all markers on load.  Removed redundancies.
